kcity/requirements.txt
paho_mqtt==1.6.1
Requests==2.29.0
